requirements.txt
atlassian-python-api==4.0.4
beautifulsoup4==4.13.4
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
cobble==0.1.4
colorama==0.4.6
Deprecated==1.2.18
dotenv==0.9.9
html2docx==1.6.0
html2text==2025.4.15
idna==3.10
jmespath==1.0.1
lxml==5.4.0
mammoth==1.9.1
mypy_extensions==1.1.0
ndjson==0.3.1
oauthlib==3.2.2
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
python-docx==1.1.2
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
soupsieve==2.7
tinycss2==1.4.0
typing_extensions==4.13.2
urllib3==2.4.0
webencodings==0.5.1
wrapt==1.17.2
